src/requirements.txt
-i https://pypi.org/simple
blinker==1.8.2; python_version >= '3.8'
cachecontrol==0.14.0; python_version >= '3.7'
cachetools==5.5.0; python_version >= '3.7'
certifi==2024.8.30; python_version >= '3.6'
cffi==1.17.0; platform_python_implementation != 'PyPy'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
cloudevents==1.9.0
cryptography==43.0.1
deprecation==2.1.0
firebase-admin==6.5.0; python_version >= '3.7'
firebase-functions==0.1.2; python_version >= '3.10'
flask==3.0.3; python_version >= '3.8'
flask-cors==5.0.0
functions-framework==3.8.1; python_version >= '3.5' and python_version < '4'
google-api-core[grpc]==2.19.2; platform_python_implementation != 'PyPy'
google-api-python-client==2.143.0; python_version >= '3.7'
google-auth==2.34.0; python_version >= '3.7'
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1; python_version >= '3.7'
google-cloud-firestore==2.18.0; python_version >= '3.7'
google-cloud-storage==2.18.2; python_version >= '3.7'
google-crc32c==1.6.0; python_version >= '3.9'
google-events==0.12.0; python_version >= '3.7'
google-resumable-media==2.7.2; python_version >= '3.7'
googleapis-common-protos==1.65.0; python_version >= '3.7'
grpcio==1.66.1
grpcio-status==1.66.1
gunicorn==23.0.0; platform_system != 'Windows'
httplib2==0.22.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
idna==3.8; python_version >= '3.6'
itsdangerous==2.2.0; python_version >= '3.8'
jinja2==3.1.4; python_version >= '3.7'
markupsafe==2.1.5; python_version >= '3.7'
msgpack==1.0.8; python_version >= '3.8'
packaging==24.1; python_version >= '3.8'
proto-plus==1.24.0; python_version >= '3.7'
protobuf==5.28.0; python_version >= '3.8'
pyasn1==0.6.0; python_version >= '3.8'
pyasn1-modules==0.4.0; python_version >= '3.8'
pycparser==2.22; python_version >= '3.8'
pyjwt[crypto]==2.9.0; python_version >= '3.8'
pyparsing==3.1.4; python_version >= '3.1'
pyyaml==6.0.2; python_version >= '3.8'
requests==2.32.3; python_version >= '3.8'
rsa==4.9; python_version >= '3.6' and python_version < '4'
typing-extensions==4.12.2; python_version >= '3.8'
uritemplate==4.1.1; python_version >= '3.6'
urllib3==2.2.2; python_version >= '3.8'
watchdog==5.0.2; python_version >= '3.9'
werkzeug==3.0.4; python_version >= '3.8'
